Inline a lot of the Vec2f functions.

+++ client.c
@@ -5091,7 +5091,7 @@ TRY
 
               vel = mul2f(sub2f(client.players[client.player].tank, old), TICKSPERSEC);
               compi = sub2f(vel, prj2f(diff, vel));
-              compj = mul2f(unit2f(diff), sqrt(fabsf((float)(SHELLVEL*SHELLVEL) - dot2f(compi, compi))));  /* fabsf is a cludge */
+              compj = mul2f(unit2f(diff), sqrt(fabs((float)(SHELLVEL*SHELLVEL) - dot2f(compi, compi))));  /* fabsf is a cludge */
 
               if ((shell = (struct Shell *)malloc(sizeof(struct Shell))) == NULL) LOGFAIL(errno)
               shell->owner = client.pills[i].owner;
